decouple storage from computation and make sure that floats are not used for computation in core code
git-svn-id: https://svn.salilab.org/imp/trunk@2105 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/core/src/DiffusionDecorator.cpp
@@ -50,7 +50,7 @@ static MillipascalSecond eta(unit::Kelvin T)
   } else {
     unsigned int i;
     for (i=1; points[i].first < T; ++i);
-    float f= ((T - points[i-1].first)
+    double f= ((T - points[i-1].first)
               /(points[i].first - points[i-1].first))
       .get_normalized_value();
     MillipascalSecond ret=
