+++ modules/core/src/MaximumSingletonScoreRestraint.cpp
@@ -23,7 +23,7 @@ MaximumSingletonScoreRestraint
 }
 
 namespace {
-  typedef internal::MinimalSet<float,
+  typedef internal::MinimalSet<double,
     SingletonContainer::ParticleIterator, std::greater<Float> > MS;
   template <class It, class F>
   MS find_minimal_set(It b, It e, F *f, unsigned int n) {
@@ -31,7 +31,7 @@ namespace {
             << std::distance(b,e) << std::endl);
     MS bestn(n);
     for (It it= b; it != e; ++it) {
-      float score
+      double score
       = IMP::internal::ContainerTraits<Particle>::evaluate(f, *it, NULL);
 
       if (bestn.can_insert(score)) {
@@ -51,7 +51,7 @@ Float MaximumSingletonScoreRestraint
   MS bestn= find_minimal_set(c_->particles_begin(),
                              c_->particles_end(), f_.get(), n_);
 
-  float score=0;
+  double score=0;
   for (unsigned int i=0; i< bestn.size(); ++i) {
     if (da) {
       IMP::internal::ContainerTraits<Particle>::evaluate(f_.get(),
